+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Following Semantic Versioning 2.
 ## Version 0.0.10 (MINOR)
 - Hide "Keywords" and "Proposed solutions" when they are empty.
 - Upgrade module's Ruby to 2.7.5
+- Fix: hide Challenges, Problems and Solutions from global search when its Participatory space is not visible
 
 ## Version 0.0.9 (PATCH)
 - Fix: show solutions only when everything is published